+++ Components/Core/Conditions/Categories/Resources.lua
@@ -87,7 +87,7 @@ ConditionCategory:RegisterCondition(1.0, "HEALTH", {
 	max = 100,
 	icon = "Interface\\Icons\\inv_alchemy_elixir_05",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = [[UnitHealth(c.Unit)/(UnitHealthMax(c.Unit)+epsilon) c.Operator c.Level]],
 	events = function(ConditionObject, c)
 		return
@@ -103,7 +103,7 @@ ConditionCategory:RegisterCondition(1.1, "HEALTH_ABS", {
 	range = 100000,
 	icon = "Interface\\Icons\\inv_alchemy_elixir_05",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = [[UnitHealth(c.Unit) c.Operator c.Level]],
 	events = function(ConditionObject, c)
 		return
@@ -534,7 +534,7 @@ if ClassicExpansionAtLeast(LE_EXPANSION_LEGION) then
 		max = 150, -- 150 with talent Voidtouched
 		icon = "Interface\\Icons\\spell_shadow_painandsuffering",
 		tcoords = CNDT.COMMON.standardtcoords,
-		maybeSecret = true,
+		deprecated = TMW.clientHasSecrets,
 		funcstr = ([[UnitPower("player", %d) c.Operator c.Level]]):format(Enum.PowerType.Insanity),
 		events = function(ConditionObject, c)
 			return
@@ -550,7 +550,7 @@ if ClassicExpansionAtLeast(LE_EXPANSION_LEGION) then
 		range = 200,
 		icon = "Interface\\Icons\\ability_warlock_demonicpower",
 		tcoords = CNDT.COMMON.standardtcoords,
-		maybeSecret = true,
+		deprecated = TMW.clientHasSecrets,
 		funcstr = ([[UnitPower("player", %d) c.Operator c.Level]]):format(Enum.PowerType.Fury),
 		events = function(ConditionObject, c)
 			return
@@ -564,7 +564,7 @@ if ClassicExpansionAtLeast(LE_EXPANSION_LEGION) then
 		range = 200,
 		icon = "Interface\\Icons\\ability_demonhunter_torment",
 		tcoords = CNDT.COMMON.standardtcoords,
-		maybeSecret = true,
+		deprecated = TMW.clientHasSecrets,
 		funcstr = ([[UnitPower("player", %d) c.Operator c.Level]]):format(Enum.PowerType.Pain),
 		events = function(ConditionObject, c)
 			return
@@ -580,7 +580,7 @@ if ClassicExpansionAtLeast(LE_EXPANSION_LEGION) then
 		max = 200,
 		icon = "Interface\\Icons\\spell_shaman_maelstromweapon",
 		tcoords = CNDT.COMMON.standardtcoords,
-		maybeSecret = true,
+		deprecated = TMW.clientHasSecrets,
 		funcstr = ([[UnitPower("player", %d) c.Operator c.Level]]):format(Enum.PowerType.Maelstrom),
 		events = function(ConditionObject, c)
 			return
@@ -596,7 +596,7 @@ if ClassicExpansionAtLeast(LE_EXPANSION_LEGION) then
 		max = 130, -- Druid tier set increases this to 130
 		icon = "Interface\\Icons\\talentspec_druid_balance",
 		tcoords = CNDT.COMMON.standardtcoords,
-		maybeSecret = true,
+		deprecated = TMW.clientHasSecrets,
 		funcstr = ([[UnitPower("player", %d) c.Operator c.Level]]):format(Enum.PowerType.LunarPower),
 		events = function(ConditionObject, c)
 			return
@@ -651,7 +651,7 @@ ConditionCategory:RegisterCondition(102.0, "DEFAULT", {
 	max = 100,
 	icon = "Interface\\Icons\\inv_alchemy_elixir_02",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = [[UnitPower(c.Unit)/(UnitPowerMax(c.Unit)+epsilon) c.Operator c.Level]],
 	events = function(ConditionObject, c)
 		return
@@ -669,7 +669,7 @@ ConditionCategory:RegisterCondition(102.1, "DEFAULT_ABS", {
 	range = 40000,
 	icon = "Interface\\Icons\\inv_alchemy_elixir_02",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = [[UnitPower(c.Unit) c.Operator c.Level]],
 	events = function(ConditionObject, c)
 		return
@@ -705,7 +705,7 @@ ConditionCategory:RegisterCondition(103.0 - offset, "MANA", {
 	max = 100,
 	icon = "Interface\\Icons\\inv_potion_126",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d)/(UnitPowerMax(c.Unit, %d)+epsilon) c.Operator c.Level]]):format(Enum.PowerType.Mana, Enum.PowerType.Mana),
 	events = function(ConditionObject, c)
 		return
@@ -721,7 +721,7 @@ ConditionCategory:RegisterCondition(103.1 - offset, "MANA_ABS", {
 	range = 40000,
 	icon = "Interface\\Icons\\inv_potion_126",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d) c.Operator c.Level]]):format(Enum.PowerType.Mana),
 	events = function(ConditionObject, c)
 		return
@@ -754,7 +754,7 @@ ConditionCategory:RegisterCondition(104.0 - offset, "ENERGY", {
 	max = 100,
 	icon = "Interface\\Icons\\inv_potion_125",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d)/(UnitPowerMax(c.Unit, %d)+epsilon) c.Operator c.Level]]):format(Enum.PowerType.Energy, Enum.PowerType.Energy),
 	events = function(ConditionObject, c)
 		return
@@ -770,7 +770,7 @@ ConditionCategory:RegisterCondition(104.1 - offset, "ENERGY_ABS", {
 	range = 200,
 	icon = "Interface\\Icons\\inv_potion_125",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d) c.Operator c.Level]]):format(Enum.PowerType.Energy),
 	events = function(ConditionObject, c)
 		return
@@ -803,7 +803,7 @@ ConditionCategory:RegisterCondition(105.0 - offset, "RAGE", {
 	max = 100,
 	icon = "Interface\\Icons\\inv_potion_120",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d)/(UnitPowerMax(c.Unit, %d)+epsilon) c.Operator c.Level]]):format(Enum.PowerType.Rage, Enum.PowerType.Rage),
 	events = function(ConditionObject, c)
 		return
@@ -819,7 +819,7 @@ ConditionCategory:RegisterCondition(105.1 - offset, "RAGE_ABS", {
 	range = 200,
 	icon = "Interface\\Icons\\inv_potion_120",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d) c.Operator c.Level]]):format(Enum.PowerType.Rage),
 	events = function(ConditionObject, c)
 		return
@@ -852,7 +852,7 @@ ConditionCategory:RegisterCondition(106.0 - offset, "FOCUS", {
 	max = 100,
 	icon = "Interface\\Icons\\inv_potion_124",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d)/(UnitPowerMax(c.Unit, %d)+epsilon) c.Operator c.Level]]):format(Enum.PowerType.Focus, Enum.PowerType.Focus),
 	events = function(ConditionObject, c)
 		return
@@ -868,7 +868,7 @@ ConditionCategory:RegisterCondition(106.1 - offset, "FOCUS_ABS", {
 	range = 200,
 	icon = "Interface\\Icons\\inv_potion_124",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d) c.Operator c.Level]]):format(Enum.PowerType.Focus),
 	events = function(ConditionObject, c)
 		return
@@ -901,7 +901,7 @@ ConditionCategory:RegisterCondition(107.0 - offset, "RUNIC_POWER", {
 	max = 100,
 	icon = "Interface\\Icons\\inv_potion_128",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d)/(UnitPowerMax(c.Unit, %d)+epsilon) c.Operator c.Level]]):format(Enum.PowerType.RunicPower, Enum.PowerType.RunicPower),
 	events = function(ConditionObject, c)
 		return
@@ -917,7 +917,7 @@ ConditionCategory:RegisterCondition(107.1 - offset, "RUNIC_POWER_ABS", {
 	range = 200,
 	icon = "Interface\\Icons\\inv_potion_128",
 	tcoords = CNDT.COMMON.standardtcoords,
-	maybeSecret = true,
+	deprecated = TMW.clientHasSecrets,
 	funcstr = ([[UnitPower(c.Unit, %d) c.Operator c.Level]]):format(Enum.PowerType.RunicPower),
 	events = function(ConditionObject, c)
 		return
@@ -956,7 +956,7 @@ if ClassicExpansionAtLeast(LE_EXPANSION_CATACLYSM) then
 		max = 100,
 		icon = "Interface\\Icons\\spell_shadow_mindflay",
 		tcoords = CNDT.COMMON.standardtcoords,
-		maybeSecret = true,
+		deprecated = TMW.clientHasSecrets,
 		funcstr = ([[UnitPower(c.Unit, %d)/(UnitPowerMax(c.Unit, %d)+epsilon) c.Operator c.Level]]):format(Enum.PowerType.Alternate, Enum.PowerType.Alternate),
 		events = function(ConditionObject, c)
 			return
@@ -975,7 +975,7 @@ if ClassicExpansionAtLeast(LE_EXPANSION_CATACLYSM) then
 		range = 200,
 		icon = "Interface\\Icons\\spell_shadow_mindflay",
 		tcoords = CNDT.COMMON.standardtcoords,
-		maybeSecret = true,
+		deprecated = TMW.clientHasSecrets,
 		funcstr = ([[UnitPower(c.Unit, %d) c.Operator c.Level]]):format(Enum.PowerType.Alternate),
 		events = function(ConditionObject, c)
 			return